# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  chat_sdk.CreateGroupRoomReq:
    properties:
      members:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - members
    - name
    type: object
...
  chat_sdk.SendFriendRequestReq:
    properties:
      message:
        example: 你好，交个朋友
        type: string
      to_user:
        example: 1001
        type: integer
    required:
    - to_user
    type: object
...
  chat_sdk.SetFriendRemarkReq:
    properties:
      friend_id:
        example: 1002
        type: integer
      remark:
        example: 老板
        type: string
    required:
    - friend_id
    type: object
...
  chat_sdk.UpdateUserAvatarReq:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
    required:
    - avatar
    type: object
...
  models.Room:
    properties:
      avatar:
        description: 房间头像
        type: string
      createdAt:
        type: string
      creator:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      creatorID:
        description: 创建者 ID
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述
        type: string
      id:
        type: integer
      isEncrypted:
        description: 是否端到端加密
        type: boolean
      isMute:
        description: 新增禁言相关字段
        type: boolean
      memberLimit:
        description: 成员上限
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      muteDailyDuration:
        description: 每日禁言持续时长（分钟）
        type: integer
      muteDailyStartTime:
        description: 每日禁言开始时间 "HH:MM"
        type: string
      muteUntil:
        description: 全员禁言截止时间（倒计时模式）
        type: string
      name:
        description: 房间名称
        type: string
      roomAccount:
        description: |-
          RoomAccount 对外房间号/群号（可用 10 位数字字符串或自定义规则），用于搜索/分享；
          不参与任何外键关联，避免再被 GORM 推断成 bigint。
        type: string
      roomUsers:
        items:
          $ref: '#/definitions/models.RoomUser'
        type: array
      type:
        description: '类型: 1-私聊 2-群聊'
        type: integer
      updatedAt:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 业务状态码
        example: 0
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 提示消息
        example: success
        type: string
    type: object
...
  service.ConversationListItemDTO:
    properties:
      avatar:
        description: 私聊：对方头像；群聊：群头像
        type: string
      conversation_id:
        type: integer
      last_message:
        $ref: '#/definitions/service.MessageDTO'
      name:
        description: 私聊：对方昵称；群聊：群名
        type: string
      room_account:
        type: string
      room_id:
        type: integer
      room_type:
        description: 1-私聊 2-群聊
        type: integer
      unread_count:
        type: integer
      updated_at:
        description: unix seconds for easy sort/render
        type: integer
      user_id:
        description: 私聊时为对方用户ID，群聊时为0
        type: integer
    type: object
...
  service.FriendApplyDTO:
    properties:
      created_at:
        type: string
      from_user:
        $ref: '#/definitions/service.UserBasicDTO'
      id:
        type: integer
      reason:
        type: string
      status:
        type: integer
    type: object
  service.GroupInfoDTO:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      creator_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      room_account:
        type: string
      updated_at:
        type: string
    type: object
...
  service.MessageDTO:
    properties:
      content:
        type: string
      created_at:
        type: string
      extra:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_encrypted:
        type: boolean
      is_system:
        type: boolean
      message_id:
        type: string
      reply_to_msg_id:
        type: integer
      room_id:
        type: integer
      sender_id:
        type: integer
      status:
        type: integer
      type:
        type: integer
      updated_at:
        type: string
    type: object
  service.MessageListItemDTO:
    properties:
      content:
        type: string
      created_at:
        type: string
      extra:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_encrypted:
        type: boolean
      is_system:
        type: boolean
      reply_to_msg_id:
        type: integer
      room_id:
        type: integer
      sender:
        $ref: '#/definitions/service.SenderDTO'
      sender_id:
        type: integer
      status:
        type: integer
      type:
        type: integer
      updated_at:
        type: string
    type: object
...
  service.RoomDTO:
    properties:
      avatar:
        type: string
      id:
        type: integer
      last_message:
        $ref: '#/definitions/service.MessageDTO'
      name:
        type: string
      room_account:
        type: string
      type:
        type: integer
      unread_count:
        type: integer
      updated_at:
        type: string
    type: object
...
  service.UserDTO:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        type: integer
      group_nickname:
        description: 我在该群里的昵称（群成员/会话列表可用）
        type: string
      id:
        type: integer
      last_active_at:
        type: string
      last_login_at:
        type: string
      nickname:
        type: string
      online_status:
        type: integer
      phone:
        type: string
      remark:
        description: 好友备注（仅在好友/私聊场景有意义）
        type: string
      room_account:
        description: 私聊房间对外号（与该好友的会话）
        type: string
      room_id:
        description: 私聊房间ID（与该好友的会话）
        type: integer
      signature:
        type: string
      uid:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:6789
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cydxin/chat-sdk/issues
  description: |-
    即时通讯 SDK 的 RESTful API 文档，包含用户管理、消息、房间、好友等模块

    ## 业务状态码说明
    | Code | 说明 |
    |------|------|
    | 0 | 成功 |
    | 10001 | 参数错误 |
    | 10002 | 用户不存在 |
    | 10003 | 密码错误（登录失败） |
    | 10004 | Token 无效 |
    | 10005 | 权限不足 |
    | 99999 | 内部错误 |

    ## HTTP 状态码说明
    - **200**: 业务请求成功（根据 response.code 判断业务状态）
    - **401**: 认证失败（未登录/Token 无效/登录失败）
    - **403**: 权限不足
    - **500**: 服务器内部错误

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
    "code": 0,
    "msg": "success",
    "data": {}
    }
    ```
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/cydxin/chat-sdk
  title: Chat SDK API
  version: "1.0"
paths:
  /friend/accept:
    post:
      consumes:
      - application/json
      description: 同意指定的好友申请
      parameters:
      - description: 申请ID
        format: int64
        in: query
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 同意好友申请
      tags:
      - 好友
...
  /friend/delete:
    post:
      consumes:
      - application/json
      description: 删除好友关系
      parameters:
      - description: 好友ID
        format: int64
        in: query
        name: friend_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除好友
      tags:
      - 好友
  /friend/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户的好友列表
      produces:
      - application/json
      responses:
        "200":
          description: 好友列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.UserDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取好友列表
      tags:
      - 好友
  /friend/pending:
    get:
      consumes:
      - application/json
      description: 获取当前用户的好友申请列表
      produces:
      - application/json
      responses:
        "200":
          description: 好友申请列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.FriendApplyDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取好友申请
      tags:
      - 好友
  /friend/reject:
    post:
      consumes:
      - application/json
      description: 拒绝指定的好友申请
      parameters:
      - description: 申请ID
        format: int64
        in: query
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 拒绝好友申请
      tags:
      - 好友
  /friend/remark:
    post:
      consumes:
      - application/json
      description: 设置当前用户对某个好友的备注（仅影响自己视角）
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetFriendRemarkReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置好友备注
      tags:
      - 好友
  /friend/request:
    post:
      consumes:
      - application/json
      description: 向目标用户发送好友申请
      parameters:
      - description: 好友申请
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SendFriendRequestReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 发送好友申请
      tags:
      - 好友
...
  /message/conversation/hide:
    post:
      consumes:
      - application/json
      description: 将当前用户某个房间的会话从消息列表隐藏（仅影响自己；新消息会自动重新展示）
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 隐藏会话
      tags:
      - 消息
  /message/conversations:
    get:
      consumes:
      - application/json
      description: 获取当前用户的会话列表（未删除的会话），包含头像、名称、room、最后一条消息、未读数
      produces:
      - application/json
      responses:
        "200":
          description: 会话列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.ConversationListItemDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取消息列表
      tags:
      - 消息
  /message/detail:
    get:
      consumes:
      - application/json
      description: 根据消息ID获取消息详情
      parameters:
      - description: 消息ID
        format: int64
        in: query
        name: message_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 消息详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.MessageDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取消息详情
      tags:
      - 消息
...
  /message/list:
    get:
      consumes:
      - application/json
      description: 分页获取房间历史消息
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 偏移量 以你要查询的ID为基准，向前查询，不传则向后
        in: query
        name: mess_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 消息列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.MessageListItemDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取房间消息
      tags:
      - 消息
  /message/recall:
    post:
      consumes:
      - application/json
      description: 撤回指定消息
      parameters:
      - description: 消息ID
        format: int64
        in: query
        name: message_id
        required: true
        type: integer
      - description: '4-撤回（会在聊天窗口留下痕迹） 5-删除（自己不可见） 6/7-双删（Sender/非Sender删除)在私聊中互相可以删除，但在群中你只能删除自己的，已经管理员进行删除 '
        format: int32
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 撤回消息
      tags:
      - 消息
...
  /room/group:
    post:
      consumes:
      - application/json
      description: 创建新的群聊房间
      parameters:
      - description: 创建参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.CreateGroupRoomReq'
      produces:
      - application/json
      responses:
        "200":
          description: 房间信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Room'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建群聊
      tags:
      - 房间
  /room/group/info:
    get:
      consumes:
      - application/json
      description: 根据 room_id 获取群聊基础信息（不含成员列表）
      parameters:
      - description: 群ID(房间ID)
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.GroupInfoDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取群基础信息
      tags:
      - 房间
  /room/group/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户参与的所有群聊（仅 Type=2）
      produces:
      - application/json
      responses:
        "200":
          description: 群聊列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.RoomDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户群聊列表
      tags:
      - 房间
...
  /room/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户参与的所有房间
      produces:
      - application/json
      responses:
        "200":
          description: 房间列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.RoomDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户房间列表
      tags:
      - 房间
  /room/member/add:
    post:
      consumes:
      - application/json
      description: 将用户添加到房间
      parameters:
      - description: 成员信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.RoomMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 添加房间成员
      tags:
      - 房间
...
  /room/member/list:
    get:
      consumes:
      - application/json
      description: 获取指定房间(群)成员列表，展示名按：好友备注 > 群昵称 > 用户昵称 > 用户名
      parameters:
      - description: 房间ID
        format: int64
        in: query
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成员列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.RoomMemberListItemDTO'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取群成员列表
      tags:
      - 房间
  /room/member/nickname:
    post:
      consumes:
      - application/json
      description: 设置当前用户在指定房间(群)中的昵称（room_user.nickname），仅影响自己视角
      parameters:
      - description: 请求参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SetMyGroupNicknameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置群昵称
      tags:
      - 房间
  /room/member/remove:
    post:
      consumes:
      - application/json
      description: 将用户从房间移除
      parameters:
      - description: 成员信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.RoomMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 移除房间成员
      tags:
      - 房间
...
  /room/private:
    post:
      consumes:
      - application/json
      description: 创建或获取两人私聊房间
      parameters:
      - description: 目标用户ID
        format: int64
        in: query
        name: target_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 房间信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Room'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建私聊
      tags:
      - 房间
  /user/avatar:
    post:
      consumes:
      - application/json
      description: 更新当前用户头像 URL
      parameters:
      - description: 头像更新
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.UpdateUserAvatarReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDTO'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户头像
      tags:
      - 用户
  /user/code/send:
    post:
      consumes:
      - application/json
      description: 发送验证码到手机号/邮箱（identifier=手机号/邮箱），purpose=register/forgot_password
      parameters:
      - description: 发送验证码请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.SendVerifyCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.SendCodeResult'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 发送验证码
      tags:
      - 用户
  /user/info:
    get:
      consumes:
      - application/json
      description: 根据 user_id 查询用户详情，如果不传 user_id 则查询当前登录用户
      parameters:
      - description: 用户ID (不传则查自己)
        format: int64
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDTO'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录并返回 token（account 支持 username/phone/email；password 或 code 二选一）
      parameters:
      - description: 登录信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录响应（token + 用户信息）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.LoginResp'
              type: object
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 用户
  /user/password:
    post:
      consumes:
      - application/json
      description: 修改当前用户密码（验证码/鉴权由调用层保证）
      parameters:
      - description: 密码修改
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/chat_sdk.UpdateUserPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改用户密码
      tags:
      - 用户
  /user/password/forgot:
    post:
      consumes:
      - application/json
      description: 通过验证码重置密码（identifier 支持 phone/email/username；推荐 phone/email）
      parameters:
      - description: 忘记密码请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.ForgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 忘记密码
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号：username + (phone/email 二选一) + password + code
      parameters:
      - description: 注册信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 用户
  /user/search:
    get:
      consumes:
      - application/json
      description: 按关键字搜索用户（username/nickname/uid），自动排除当前用户
      parameters:
      - description: 搜索关键字
        in: query
        name: keyword
        type: string
      - description: 返回条数
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.UserDTO'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 搜索用户
      tags:
      - 用户
  /user/update:
    post:
      consumes:
      - application/json
      description: 更新当前用户资料（昵称/签名/性别等）
      parameters:
      - description: 更新信息（可选字段）
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserDTO'
              type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
  QueryToken:
    description: 用于 WebSocket 等无法传 header 的场景
    in: query
    name: token
    type: apiKey
swagger: "2.0"